Moved wrapper classes into .wrappers sub-module.
Also move the array_index-related methods to the base classes since they can be
effectively be the same for all implementations
