+++ astropy/wcs/wcsapi/__init__.py
@@ -1,5 +1,5 @@
-from .low_level_api import *
-from .high_level_api import *
-from .high_level_wcs_wrapper import *
-from .utils import *
-from .sliced_low_level_wcs import *
+from .low_level_api import *  # noqa
+from .high_level_api import *  # noqa
+from .high_level_wcs_wrapper import *  # noqa
+from .utils import *  # noqa
+from .wrappers import *  # noqa
