+++ astropy/wcs/wcsapi/tests/test_high_level_api.py
@@ -16,13 +16,7 @@ class DoubleLowLevelWCS(BaseLowLevelWCS):
     def pixel_to_world_values(self, *pixel_arrays):
         return [np.asarray(pix) * 2 for pix in pixel_arrays]
 
-    def array_index_to_world_values(self, *index_arrays):
-        return [np.asarray(pix) * 2 for pix in index_arrays]
-
     def world_to_pixel_values(self, *world_arrays):
-        return [np.asarray(world) / 2 for world in world_arrays]
-
-    def world_to_array_index_values(self, *world_arrays):
         return [np.asarray(world) / 2 for world in world_arrays]
