+++ astropy/wcs/wcsapi/wrappers/__init__.py
@@ -1,0 +1,1 @@
+from .sliced_wcs import *  # noqa
